Update PHPCS Version
Update to the latest version of PHPCS and fix the errors that it
detected.

Note that create_function is replaced by an anonymous function that
uses an eval. This is bad, but the create_function documents say
it uses an eval internally, so it's a small step forward and gets
rid of the error about create_function being deprecated.

Another pass should get rid of the evals.