[Core] Remove main.php file
Instrument loading (now done with the instruments module) and the API (now done
with the API module) were the last 2 stragglers requiring the legacy main.php
file to exist. The file can be removed as of the 22.0 release, as the .htaccess
file no longer references it at all and it should not be used.