[Core] Require correct types for lorisform->addSelect
Since the type is currently only enforced by PHPDocs and not
PHP, incorrect types can be passed by addSelect. This enforces
the types from the PHP side of things.

(Currently when a string is passed instread of an array in some
instruments, the array_merge can generated a PHP warning which
breaks endpoints which use an instrument in a JSON context.)