[tests] Speed up tests by caching Raisinbread in Docker image
This increases the speed of our API tests by removing the sourcing
of raisinbread from the test job and having a separate job conditionally
build a docker image containing raisinbread to push to ghcr.io. The
tests then use that database Docker image instead of duplicating the
work of sourcing the entire database with every run. The DB image is
only re-built when one of the files that it depends on is modified.

(The docker-compose tests can not be sped up this way since it depends
on GitHub Actions calculating the image to use.)