[jsx/Modal] Add onCancel option to Modal
Currently, there is no way for a component to handle a Modal being
cancelled if the state is managed by the component instead of the
modal.

This adds an onCancel so that components which use a modal and store the
open/closed state can close the modal without saving if cancelled.