[TypeScript] Add StaticDataTable type definition
This adds a typescript module definition for StaticDataTable, so that
it can be imported from typescript in strict mode without triggering
errors.

It also adds a new parameter to the "getFormattedColumn" callback
to indicate the current index of the column. (This should technically
be a different PR, but it's easier to do it at the same time rather
than need to modify the definition after.)