[dataquery] Add API endpoint to re-load previously run queries
This implements the endpoint that is listed as "aspirational" in the
dataquery api schema to load the results of a previously run query.

It was originally left out for fear of performance issues but with
this implementation performance seems to be comparable to without
it.

This does not implement a frontend to reload query results, but
the API functionality is useful for eg api callbacks that want
to process the results of a query without the need to re-run
it (and the possibility to get different results.)