[Core] Add performance profiler logger
This adds a new Profiler class that can be used for logging performance
bottlenecks in LORIS. To use it, add a call to Profiler::checkpoint("message")
and if the profiler is enabled in the configuration module (default, disabled)
it will print to the error logs a timestamp, how long the request has taken
so far, the message, max memory usage, and any new memory usage since the last
checkpoint call.

This can be used to debug where requests are slowing down. The checkpoints added
in this PR are the places that I used to optimize the dataquery module.